backport change https://github.com/nmoinvaz/minizip/commit/4e3541b6d9f93dbab823d9016a8cb1fde55fab8e
Allow reading > 64k at a time on platforms where zlib supports this.

zlib uses uInt (`unsigned int`) for avail_in and avail_out.
This type can be limited to 64k and still conform to ANSI C,
but on most platforms it's bigger.

Remove some unnecessarily-narrow casting and only enforce
the actual limit (UINT_MAX), which might be 64k or might be larger.